requirements.txt
playwright==1.40.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
fake-useragent==1.4.0
sqlite3
dataclasses-json==0.6.3